Split getWebauthnAssertionActions into separate file
