+++ src/login/redux/actions/postResetPasswordActions.js
@@ -8,8 +8,6 @@ export const POST_RESET_PASSWORD_VERIFY_EMAIL_SUCCESS = "POST_RESET_PASSWORD_VER
 export const POST_RESET_PASSWORD_EXTRA_SECURITY_PHONE = "POST_RESET_PASSWORD_EXTRA_SECURITY_PHONE";
 export const POST_RESET_PASSWORD_EXTRA_SECURITY_PHONE_FAIL = "POST_RESET_PASSWORD_EXTRA_SECURITY_PHONE_FAIL";
 export const SAVE_PHONE_CODE = "SAVE_PHONE_CODE";
-export const GET_WEBAUTHN_ASSERTION = "GET_WEBAUTHN_ASSERTION";
-export const CANCLE_WEBAUTHN_ASSERTION = "CANCLE_WEBAUTHN_ASSERTION";
 
 export function postEmailLink(email) {
   return {
@@ -82,24 +80,6 @@ export function savePhoneCode(code) {
     type: SAVE_PHONE_CODE,
     payload: {
       phone_code: code
-    }
-  };
-}
-
-export function getWebauthnAssertion(assertion) {
-  return {
-    type: GET_WEBAUTHN_ASSERTION,
-    payload: {
-      webauthn_assertion: assertion
-    }
-  };
-}
-
-export function cancleWebauthnAssertion() {
-  return {
-    type: CANCLE_WEBAUTHN_ASSERTION,
-    payload: {
-      webauthn_assertion: undefined
     }
   };
 }
